Replace NSWindowController+SwinjectSpec with NSWindowController+SwinjectTests.
